backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.3.0
httpx>=0.24.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
PyJWT>=2.8.0
openai>=1.0.0
